requirements.txt
pydantic-settings>=2.9.1
pika>=1.3.2
msgpack>=1.1.0
uvicorn>=0.34.2
fastapi>=0.115.12
python-jose>=3.4.0
pydantic>=2.11.4
dotenv>=0.9.9
python-dotenv>=1.1.0
sqlalchemy>=2.0.40
pwdlib>=0.2.1
APScheduler>=3.11.0